Events can now be saved to gdf files.
Former-commit-id: 0ff84d9d784e81a26a45d88ea258a704dc2c4dc9
Former-commit-id: b620cd12354ec0ed81a6d04fbb8df205c2632d45

+++ App/src/DataFile/eventtable.h
@@ -52,8 +52,8 @@ public:
 private:
 	std::vector<std::string> label;
 	std::vector<int> type;
-	std::vector<qlonglong> position;
-	std::vector<qlonglong> duration;
+	std::vector<int> position;
+	std::vector<int> duration;
 	std::vector<int> channel;
 	std::vector<std::string> description;
 };
